+++ website/core/Footer.js
@@ -45,6 +45,9 @@ class Footer extends React.Component {
           </div>
           <div>
             <h5>Community</h5>
+            <a href="https://crwd.in/react-leaflet">
+              Help translate this website
+            </a>
             <a href="https://stackoverflow.com/questions/tagged/react-leaflet">
               Stack Overflow
             </a>
@@ -68,6 +71,9 @@ class Footer extends React.Component {
             <a href="https://dns.js.org/">JS.org for the domain name</a>
             <a href="https://community.algolia.com/docsearch/">
               Algolia for search
+            </a>
+            <a href="https://crowdin.com/page/open-source-project-setup-request">
+              Crowdin for translations
             </a>
           </div>
         </section>